changed around game componets
