+++ client/good-games-ui/src/views/Game.js
@@ -1,5 +1,6 @@
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
+import Reviews from "../components/game_components/Reviews";
 
 const convert = require('xml-js');
 
@@ -48,16 +49,9 @@ const convertObj = (data) => {
     return newGame;
 }
 
-const REVIEW_DEFAULT = {
-    userId: 0,
-    gameId: 0,
-    text: "",
-    rating: 1,
-}
-
 function Game() {
     const [game, setGame] = useState({});
-    const [review, setReview] = useState(REVIEW_DEFAULT);
+    
     const [isLoading, setIsLoading] = useState(true);
 
     const game_url = 'http://localhost:8080/api/game/bggId';
@@ -95,6 +89,7 @@ function Game() {
                     if (data) {
                         gameDetails.game = data;
                         
+                        // fetch request for reviews
                         fetch(`${review_url}/game/${data.gameId}`)
                         .then(response => {
                             if (response.status === 200) {
@@ -111,6 +106,8 @@ function Game() {
                             console.log(error);
                             setGame(gameDetails);
                         })
+                    } else {
+                        setGame(gameDetails);
                     }
                 })
                 .catch(error => {
@@ -125,31 +122,6 @@ function Game() {
         }
     }, [id])
 
-    const handleChange = (event) => {
-        const newReview = {...review};
-
-        if(event.target.type === 'checkbox'){
-            newReview[event.target.name] = event.target.checked;
-        }else{
-            newReview[event.target.name] = event.target.value;
-        }
-
-        setReview(newReview);
-    }
-
-    const handleValueChange = (event) => {
-        const newReview = {...review};
-        newReview.rating = parseInt(event.target.value);
-        setReview(newReview);
-    }
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-
-
-        console.log(event)
-    }
-
     const renderPublishers = () => {
         if (game.publishers.length > 0) {
             const maxGet = game.publishers.length > 5 ? 5 : game.publishers.length;
@@ -160,48 +132,7 @@ function Game() {
         }
     }
 
-    const renderScores = () => {
-        return (
-            ""
-        )
-    }
-
-    const renderAvgScores = () => {
-        const totalScore = game.game.reviews.reduce((prev, curr) => curr.rating + prev , 0)
-        return totalScore / game.game.reviews.length;
-    }
-
-    const renderTotalReviews = () => {
-        return game.game.reviews.length;
-    }
-
-    const renderReviews = () => {
-        return game.game.reviews.map((review, index) => (
-            <div className="card col-12 mb-2" key={index}>
-                <div className="card-body">
-                    <h5 className="card-title">{"NOOOOO!"}</h5>
-                    <h6 className="card-subtitle mb-2 text-body-secondary">{review.rating}</h6>
-                    <p className="card-body">{review.text}</p>
-                </div>
-            </div>
-        ))
-    }
-
-    const renderRadios = () => {
-        return [...Array(10).keys()].map(key => (
-            <div key={key} className="form-check form-check-inline">
-                <input
-                className="form-check-input"
-                type="radio"
-                name="rating"
-                id={`star-${key + 1}`}
-                checked={review.rating === key + 1}
-                onChange={handleValueChange}
-                value={key + 1}/>
-                <label className="form-check-label">{`${key + 1}`}</label>
-            </div>
-        ))
-    }
+    
 
     return (
         <main className="container mt-4">
@@ -245,69 +176,7 @@ function Game() {
                         </div>
                     </section>
                     {game.game ? (
-                        <>
-                            <section className="card p-4 mt-3">
-                                <div className="row">
-                                    <div className="col-12">
-                                        <h3 className="card-title">Ratings</h3>
-                                        <p>Total Score: {game.game.reviews ? renderAvgScores() : ""}</p>
-                                        {renderScores()}
-                                    </div>
-                                </div>
-                            </section>
-                            <section className="card p-4 mt-3 mb-3">
-                                <div className="row">
-                                    <div className="col-10">
-                                        <h3 className="card-title">Reviews</h3>
-                                        <p>Total Reviews: {game.game.reviews ? renderTotalReviews() : ""}</p>
-                                    </div>
-                                    <div className="col-2">
-                                        <button className="btn btn-primary" data-bs-toggle="modal" data-bs-target="#newReviewModal">
-                                            Create Review
-                                        </button>
-
-                                        <div className="modal fade" id="newReviewModal" tabIndex={-1} aria-labelledby="newReviewModalLabel" aria-hidden="true">
-                                            <div className="modal-dialog">
-                                                <div className="modal-content">
-                                                    <form onSubmit={handleSubmit}>
-                                                        <div className="modal-header">
-                                                            <h1 className="modal-title fs-5" id="newReviewModalLabel">New Review</h1>
-                                                            <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                                                        </div>
-                                                        <div className="modal-body">
-                                                            <fieldset className="form-group">
-                                                                <label htmlFor="text">Review</label>
-                                                                <textarea
-                                                                id="text"
-                                                                name="text"
-                                                                className="form form-control"
-                                                                value={review.text}
-                                                                onChange={handleChange}
-                                                                />
-                                                            </fieldset>
-                                                            <div className="mt-3">
-                                                                {renderRadios()}
-                                                            </div>
-                                                            
-                                                        </div>
-                                                        <div className="modal-footer">
-                                                            <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                                                            <button type="submit" className="btn btn-primary" data-bs-dismiss="modal">Save Review</button>
-                                                        </div>
-                                                    </form>
-                                                </div>
-                                            </div>
-                                        </div>
-                                    </div>
-
-                                    <div className="col-12">
-                                        <div className="row">
-                                            {renderReviews()}
-                                        </div>
-                                    </div>
-                                </div>
-                            </section>
-                        </>
+                        <Reviews reviews={game.game.reviews}/>
                         ) : (
                             <section className="card p-4 mt-3 mb-3">
                                 <div className="text-center">
@@ -318,8 +187,6 @@ function Game() {
                     
                 </>
             )}
-            
-            
         </main>
     )
 }
